refactor: 추상화된 목록 조회기능을 통해 회원목록 조회하도록 변경
