+++ src/main/java/core/jdbc/JdbcTemplate.java
@@ -1,5 +1,9 @@
 package core.jdbc;
 
+import core.jdbc.exception.NotExistsResultException;
+import core.jdbc.exception.NotUniqueResultException;
+import core.jdbc.exception.QueryExecuteFailException;
+
 import java.sql.Connection;
 import java.sql.PreparedStatement;
 import java.sql.ResultSet;
@@ -8,12 +12,14 @@ import java.util.ArrayList;
 import java.util.List;
 
 public class JdbcTemplate {
+    private static final String QUERY_EXECUTE_FAIL_EXCEPTION_MESSAGE_PREFIX = "Query 수행중 에러 발생 - ";
+
     public void update(PreparedStatementCreator preparedStatementCreator) {
         try (Connection connection = ConnectionManager.getConnection();
              PreparedStatement preparedStatement = preparedStatementCreator.createPreparedStatement(connection)) {
             preparedStatement.executeUpdate();
         } catch (SQLException e) {
-            throw new QueryExecuteFailException("Query 수행중 에러 발생 - " + e.getCause().getMessage(), e);
+            throw new QueryExecuteFailException(QUERY_EXECUTE_FAIL_EXCEPTION_MESSAGE_PREFIX + e.getCause().getMessage(), e);
         }
     }
 
@@ -22,18 +28,18 @@ public class JdbcTemplate {
              PreparedStatement preparedStatement = preparedStatementCreator.createPreparedStatement(connection);
              ResultSet resultSet = preparedStatement.executeQuery()) {
             if (!resultSet.next()) {
-                throw new RuntimeException("Query 결과가 존재하지 않습니다.");
+                throw new NotExistsResultException("Query 결과가 존재하지 않습니다.");
             }
 
             T object = rowMapper.mapRow(resultSet);
 
             if (resultSet.next()) {
-                throw new RuntimeException("Query 결과가 1개보다 많습니다.");
+                throw new NotUniqueResultException("Query 결과가 1개보다 많습니다.");
             }
 
             return object;
         } catch (SQLException e) {
-            throw new QueryExecuteFailException("Query 수행중 에러 발생 - " + e.getCause().getMessage(), e);
+            throw new QueryExecuteFailException(QUERY_EXECUTE_FAIL_EXCEPTION_MESSAGE_PREFIX + e.getCause().getMessage(), e);
         }
     }
 
@@ -49,7 +55,7 @@ public class JdbcTemplate {
 
             return result;
         } catch (SQLException e) {
-            throw new QueryExecuteFailException("Query 수행중 에러 발생 - " + e.getCause().getMessage(), e);
+            throw new QueryExecuteFailException(QUERY_EXECUTE_FAIL_EXCEPTION_MESSAGE_PREFIX + e.getCause().getMessage(), e);
         }
     }
 }
