+++ src/main/java/core/jdbc/exception/NotExistsResultException.java
@@ -1,0 +1,7 @@
+package core.jdbc.exception;
+
+public class NotExistsResultException extends RuntimeException {
+    public NotExistsResultException(String message) {
+        super(message);
+    }
+}
