+++ src/main/java/core/jdbc/exception/NotUniqueResultException.java
@@ -1,0 +1,7 @@
+package core.jdbc.exception;
+
+public class NotUniqueResultException extends RuntimeException {
+    public NotUniqueResultException(String message) {
+        super(message);
+    }
+}
