+++ src/main/java/core/jdbc/exception/QueryExecuteFailException.java
@@ -1,4 +1,4 @@
-package core.jdbc;
+package core.jdbc.exception;
 
 public class QueryExecuteFailException extends RuntimeException {
     public QueryExecuteFailException(String message, Throwable cause) {
